+++ examples/function_call.py
@@ -85,7 +85,7 @@ fpclient.recordings.create(
         all_messages=all_messages,
         inputs=input_variables,
         session_info=session.session_info,
-        prompt_info=formatted_prompt.prompt_info,
+        prompt_version_info=formatted_prompt.prompt_info,
         call_info=call_info,
         response_info=response_info
     )