+++ examples/test_run_history_example.py
@@ -110,7 +110,7 @@ def call_and_record(
             all_messages=all_messages,
             session_info=session_info,
             inputs=input_variables,
-            prompt_info=formatted_prompt.prompt_info,
+            prompt_version_info=formatted_prompt.prompt_info,
             call_info=call_info,
             response_info=response_info,
             trace_info=trace_info,