+++ examples/test_run_trace_example.py
@@ -101,7 +101,7 @@ for test_case in test_run.get_trace_test_cases():
             all_messages=final_messages,
             session_info=session.session_info,
             inputs={"name": test_case.input},
-            prompt_info=formatted_prompt.prompt_info,
+            prompt_version_info=formatted_prompt.prompt_info,
             call_info=call_info,
             response_info=response_info,
             test_run_info=test_run_info,