+++ examples/test_run_with_media.py
@@ -78,7 +78,7 @@ for test_case in test_run.test_cases:
             session_info=session.session_info,
             inputs=test_case.variables,
             media_inputs=media_inputs,
-            prompt_info=formatted_prompt.prompt_info,
+            prompt_version_info=formatted_prompt.prompt_info,
             call_info=call_info,
             response_info=response_info,
             test_run_info=test_run_info,