+++ tests/integration/test_anthropic_media.py
@@ -62,7 +62,7 @@ class TestAnthropicMedia(unittest.TestCase):
                 session_info=self.freeplay_client.sessions.create().session_info,
                 inputs=input_variables,
                 media_inputs=media_inputs,
-                prompt_info=formatted_prompt.prompt_info,
+                prompt_version_info=formatted_prompt.prompt_info,
                 call_info=CallInfo.from_prompt_info(formatted_prompt.prompt_info, time.time(), time.time() + 1),
             )
         )
@@ -110,7 +110,7 @@ class TestAnthropicMedia(unittest.TestCase):
                 session_info=self.freeplay_client.sessions.create().session_info,
                 inputs=input_variables,
                 media_inputs=media_inputs,
-                prompt_info=formatted_prompt.prompt_info,
+                prompt_version_info=formatted_prompt.prompt_info,
                 call_info=CallInfo.from_prompt_info(formatted_prompt.prompt_info, time.time(), time.time() + 1),
             )
         )