+++ tests/integration/test_bedrock_converse.py
@@ -53,7 +53,7 @@ class TestBedrockConverse(unittest.TestCase):
                 ],
                 session_info=self.freeplay_client.sessions.create().session_info,
                 inputs=input_variables,
-                prompt_info=formatted_prompt.prompt_info,
+                prompt_version_info=formatted_prompt.prompt_info,
                 call_info=CallInfo.from_prompt_info(
                     formatted_prompt.prompt_info, start_time, time()
                 ),
@@ -93,7 +93,7 @@ class TestBedrockConverse(unittest.TestCase):
                 ],
                 session_info=self.freeplay_client.sessions.create().session_info,
                 inputs=input_variables,
-                prompt_info=formatted_prompt.prompt_info,
+                prompt_version_info=formatted_prompt.prompt_info,
                 call_info=CallInfo.from_prompt_info(
                     formatted_prompt.prompt_info, start_time, time()
                 ),