+++ tests/integration/test_gemini_media.py
@@ -67,7 +67,7 @@ class TestGeminiMedia(unittest.TestCase):
                 session_info=self.freeplay_client.sessions.create().session_info,
                 inputs=input_variables,
                 media_inputs=media_inputs,
-                prompt_info=formatted_prompt.prompt_info,
+                prompt_version_info=formatted_prompt.prompt_info,
                 call_info=CallInfo.from_prompt_info(formatted_prompt.prompt_info, time.time(), time.time() + 1),
             )
         )
@@ -120,7 +120,7 @@ class TestGeminiMedia(unittest.TestCase):
                 session_info=self.freeplay_client.sessions.create().session_info,
                 inputs=input_variables,
                 media_inputs=media_inputs,
-                prompt_info=formatted_prompt.prompt_info,
+                prompt_version_info=formatted_prompt.prompt_info,
                 call_info=CallInfo.from_prompt_info(formatted_prompt.prompt_info, time.time(), time.time() + 1),
             )
         )
@@ -173,7 +173,7 @@ class TestGeminiMedia(unittest.TestCase):
                 session_info=self.freeplay_client.sessions.create().session_info,
                 inputs=input_variables,
                 media_inputs=media_inputs,
-                prompt_info=formatted_prompt.prompt_info,
+                prompt_version_info=formatted_prompt.prompt_info,
                 call_info=CallInfo.from_prompt_info(formatted_prompt.prompt_info, time.time(), time.time() + 1),
             )
         )